+++ src/settings/VersionHistory/VersionHistory.js
@@ -9,7 +9,6 @@ import {
   Row,
   Col,
   Select,
-  Label,
   Layout,
   LoadingPane,
 } from '@folio/stripes/components';
@@ -35,7 +34,7 @@ const VersionHistory = () => {
 
   const pageSizeSettings = settings?.find(setting => setting.key === VERSION_HISTORY_PAGE_SIZE_SETTING);
 
-  const pageSizeOptions = [10, 25, 50, 100].map(value => ({ label: value, value }));
+  const pageSizeOptions = [10, 25, 50, 75, 100].map(value => ({ label: value, value }));
 
   const handleSubmit = async formValues => {
     try {
@@ -82,15 +81,11 @@ const VersionHistory = () => {
           onSubmit={handleSubmit}
         >
           <Row>
-            <Col xs={12}>
-              <Label htmlFor={fieldNames.PAGE_SIZE}>
-                {intl.formatMessage({ id: 'ui-marc-authorities.settings.versionHistory.field.pageSize' })}
-              </Label>
-            </Col>
-            <Col xs={4}>
+            <Col xs={6}>
               <Field
                 id={fieldNames.PAGE_SIZE}
                 name={fieldNames.PAGE_SIZE}
+                label={intl.formatMessage({ id: 'ui-marc-authorities.settings.versionHistory.field.pageSize' })}
                 dataOptions={pageSizeOptions}
                 component={Select}
                 parse={v => parseInt(v, 10)}