+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Change history for ui-marc-authorities
 
-## [7.0.0] (IN PROGRESS)
+## [7.0.0] IN PROGRESS
 
-- [UIMARCAUTH-445](https://issues.folio.org/browse/UIMARCAUTH-445) MARC authority > View Source > Display Version History pane with an empty Version History component.
+- [UIMARCAUTH-446](https://issues.folio.org/browse/UIMARCAUTH-446) *BREAKING* Upgrade @folio/stripes-* dependencies.
+- [UIMARCAUTH-447](https://issues.folio.org/browse/UIMARCAUTH-447) *BREAKING* Upgrade react-intl to v7.
 - [UIMARCAUTH-444](https://issues.folio.org/browse/UIMARCAUTH-444) *BREAKING* Create MARC authority settings to configure number of cards in version history.
+- [UIMARCAUTH-445](https://issues.folio.org/browse/UIMARCAUTH-445) MARC authority > View Source > Display Version History pane with an empty Version History component.
 
 ## [6.0.1](https://github.com/folio-org/ui-marc-authorities/tree/v6.0.1) (2024-12-13)